changed price to BigDecimal

+++ src/main/java/kz/kakimzhanova/delivery/dao/OrderDao.java
@@ -3,7 +3,9 @@ package kz.kakimzhanova.delivery.dao;
 import kz.kakimzhanova.delivery.entity.Order;
 import kz.kakimzhanova.delivery.exception.DaoException;
 
+import java.math.BigDecimal;
+
 public interface OrderDao extends BaseDao<Integer, Order>{
     Order create(String login) throws DaoException;
-    boolean updateTotalCost(int orderId, float totalCost) throws DaoException;
+    boolean updateTotalCost(int orderId, BigDecimal totalCost) throws DaoException;
 }
